# sechuv_chve/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SecHuv",
    "description": "Security Hub for Human-vulnerabilities.",
    "contact": {
      "email": "tsubasa.umeuchi@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/case": {
      "get": {
        "tags": [
          "general"
        ],
        "summary": "指定した事例を取得",
        "parameters": [
          {
            "name": "uuid",
            "in": "query",
            "description": "取得する事例のUUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "description": "取得する事例のメディア",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "is_valid",
            "in": "query",
            "description": "正当な事例であるかどうか",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Case"
                }
              }
            }
          },
          "400": {
            "description": "Parameter error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "general"
        ],
        "summary": "指定した事例を削除",
        "parameters": [
          {
            "name": "uuid",
            "in": "query",
            "description": "削除する事例のUUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "description": "取得する事例のメディア",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "is_valid",
            "in": "query",
            "description": "正当な事例であるかどうか",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "uuid": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parameter error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/web/case": {
      "get": {
        "tags": [
          "web"
        ],
        "summary": "Webに関する事例を取得",
        "parameters": [
          {
            "name": "length",
            "in": "query",
            "description": "取得する件数",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WebCase"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "web"
        ],
        "summary": "Webに関する事例を投稿",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebCasePost"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "uuid": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "post_date": {
                      "type": "string",
                      "format": "yyyymmddhhmmss"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/web/valid": {
      "get": {
        "tags": [
          "web"
        ],
        "summary": "Webに関する正当なコンテンツを取得",
        "parameters": [
          {
            "name": "length",
            "in": "query",
            "description": "取得する件数",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WebValidCase"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "web"
        ],
        "summary": "Webに関する正当なコンテンツを取得",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "uuid": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "post_date": {
                      "type": "string",
                      "format": "yyyymmddhhmmss"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/web/check": {
      "post": {
        "tags": [
          "web"
        ],
        "summary": "Webコンテンツの危険度をチェック",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "vulns": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Vulnerability"
                      }
                    },
                    "samples": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/WebCase"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/mail/case": {
      "get": {
        "tags": [
          "mail"
        ],
        "summary": "メールに関する事例を取得",
        "parameters": [
          {
            "name": "length",
            "in": "query",
            "description": "取得する件数",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MailCase"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "mail"
        ],
        "summary": "メールに関する事例を投稿",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MailCasePost"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "uuid": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "post_date": {
                      "type": "string",
                      "format": "yyyymmddhhmmss"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/mail/valid": {
      "get": {
        "tags": [
          "mail"
        ],
        "summary": "メールに関する正当なコンテンツを取得",
        "parameters": [
          {
            "name": "length",
            "in": "query",
            "description": "取得する件数",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MailValidCase"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "mail"
        ],
        "summary": "メールに関する正当なコンテンツを投稿",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MailSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "uuid": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "post_date": {
                      "type": "string",
                      "format": "yyyymmddhhmmss"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/mail/check": {
      "post": {
        "tags": [
          "mail"
        ],
        "summary": "メールコンテンツの危険度をチェック",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MailSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "vulns": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Vulnerability"
                      }
                    },
                    "samples": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MailCase"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/other/case": {
      "get": {
        "tags": [
          "other"
        ],
        "summary": "その他の事例を取得",
        "parameters": [
          {
            "name": "length",
            "in": "query",
            "description": "取得する件数",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OtherCase"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "other"
        ],
        "summary": "その他の事例を投稿",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtherCasePost"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "uuid": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "post_date": {
                      "type": "string",
                      "format": "yyyymmddhhmmss"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/other/valid": {
      "get": {
        "tags": [
          "other"
        ],
        "summary": "その他の正当なコンテンツを取得",
        "parameters": [
          {
            "name": "length",
            "in": "query",
            "description": "取得する件数",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OtherValidCase"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "other"
        ],
        "summary": "その他の正当なコンテンツを投稿",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtherSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "uuid": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "post_date": {
                      "type": "string",
                      "format": "yyyymmddhhmmss"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Case": {
        "type": "object",
        "oneOf": [
          {
            "$ref": "#/components/schemas/WebCase"
          },
          {
            "$ref": "#/components/schemas/MailCase"
          },
          {
            "$ref": "#/components/schemas/OtherCase"
          }
        ]
      },
      "WebCase": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "format": "uuid"
          },
          "post_date": {
            "type": "string",
            "format": "yyyymmddhhmmss"
          },
          "vulntypes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "spec": {
            "$ref": "#/components/schemas/WebSpec"
          }
        }
      },
      "WebCasePost": {
        "type": "object",
        "properties": {
          "vulntypes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "spec": {
            "$ref": "#/components/schemas/WebSpec"
          }
        }
      },
      "WebSpec": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "raw_body": {
            "type": "string"
          },
          "screenshot": {
            "type": "string"
          }
        }
      },
      "WebValidCase": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "format": "uuid"
          },
          "post_date": {
            "type": "string",
            "format": "yyyymmddhhmmss"
          },
          "spec": {
            "$ref": "#/components/schemas/WebSpec"
          }
        }
      },
...
      "Vulnerability": {
        "type": "object",
        "properties": {
          "vulntype": {
            "type": "string"
          },
          "vulnname": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      }
    }
  }
}